major refactor
- put everything in a single file
